first working version

+++ vec.c
@@ -4,8 +4,8 @@
 GEN_VEC(u8, u8)
 GEN_VEC_COMPARE(u8, u8)
 
-GEN_VEC(u16, u16)
-GEN_VEC_COMPARE(u16, u16)
+GEN_VEC(u32, u32)
+GEN_VEC_COMPARE(u32, u32)
 
 void append_string(u8Vec *vector, char *string) {
     u8 c;
